requirements.txt
selenium
bs4
pytest
datetime
os
pymongo
python-dotenv
bson